requirements.txt
flask>=2.0.0
flask-sqlalchemy>=2.5.0
python-dotenv>=0.19.0
Pillow>=9.0.0
efficientnet-pytorch>=0.7.1
torch>=1.13.0
torchvision>=0.14.0
opencv-python-headless>=4.5.0
face-recognition>=1.3.0
reverse-geocoder>=1.5.1
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
numpy>=1.21.0

